Update phase 3 ranges and add more sources for required chars.
A suggestion by Behdad said that complex scripts should probably include
ZWJ/ZWNJ and dotted circle, even though harfbuzz can usually handle text
without these in the font.  This adds a list of complex scripts, gleaned
by examining the harfbuzz source in hb-ot-shape-complex-private.hh, and
adds code that checks if the font's script is complex and if so adds these
characters to the required characters.

A suggestion from Jungshik said that we should probably look at cldr
data to get punctuation characters.  This adds a tool that extracts some
punctuation from a few fields of the punctuation data.  This data is
then hard-coded into the noto_data.py file, once I figure out how best to
keep the required character data stable I might change this.

Another suggestion from Jungshik led to removing ASCII_DIGITS from the
scripts I'd added them to previously; we expect these to not interact
typographically so falling back to LGC is probably ok-- the punctuation
is for cases where we're not so sure.

A few more data sets were converted to the format used by parse_int_ranges,
but the data is still inconsistently formatted.

Characters supported in Phase2 Indic were added to the requirements for
Phase 3, mostly to quiet the reporting on removed characters from some of
the tooling.  We anticipate keeping these characters.