Tweak notoconfig.
It's really obscure when notoconfig fails because people don't have a
file set up.  Add some documentation, print a warning if it is not set
up, default 'noto_tools' to the root of the nototools repo containing
the file, and provide a getter that fails if a key is not defined
(except for noto_tools).