Update swat_release_font_names use of glyphsLib.
This uses the Reader/Writer version of glyphsLib that reads/writes
glyphs data preserving quotes.  Upstream is implementing changes that
will break this API, so for the time being the only way to swat glyphs
data without normalizing it first is to use the right commits of
glyphsLib and this.

Also updates requirements.txt to add a dependency on glyphsLib.