Add tool to swat glyphs file version info.
Also updates the tool that swats font names a bit since it is
used by the new tool

The new tool hacks the glyphs files by replacing lines. Unlike
the other tool it doesn't parse the tree since it doesn't need
to understand much about the structure, so it does not have a
dependency on glyphsLib.